pyproject.toml
[tool.poetry]
name = "check-open-source-releases"
version = "0.1.0"
description = ""
authors = ["Paul Luna <paulluna0215@gmail.com>"]
readme = "README.md"
packages = [{include = "check_open_source_releases"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
pygithub = "^1.59.1"

[tool.poetry.scripts]
main = "scripts:main"
add_label = "scripts:add_label"
create_issue = "scripts:create_issue"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
